+++ config/routes.rb
@@ -287,6 +287,7 @@ Rails.application.routes.draw do
     member do
       get :profile_picture
       get :posts
+      get :tools
     end
   end